Adding docstrings to tests, moving tests to a separate folder
