+++ pennylane/circuit_graph.py
@@ -118,8 +118,8 @@ class CircuitGraph:
                 self._graph.add_edge(wire[i - 1], wire[i])
 
     def create_hash_argument(self):
-        """Create a unique hash based on the operation queue, observables queue and variable
-        deps.
+        """Create a unique hash based on the operations and observables in the circuit graph and the index of the variables
+        used by them.
         """
         key_to_be_hashed = ''
 
@@ -141,7 +141,7 @@ class CircuitGraph:
 
             key_to_be_hashed += str(op.wires)
 
-        # Adding a distinct separating string that could not occur by any combining the
+        # Adding a distinct separating string that could not occur by any combination of the
         # name of the operation and wires
         key_to_be_hashed += '|||'
 
@@ -165,6 +165,8 @@ class CircuitGraph:
 
     @property
     def circuit_hash(self):
+        """Creating a hash for the circuit graph based on the argument generated by create_hash_argument.
+        """
         return hash(self.create_hash_argument())
 
     @property
