+++ tests/qnodes/test_qnode_base.py
@@ -812,645 +812,6 @@ class TestQNodeEvaluate:
         res = node(0.432, 0.12)
         assert res.shape == (10,)
 
-class TestQNodeCircuitHashIntegration:
-    """Test for the circuit hash that is being created for a QNode during evaluation (inside of _construct)"""
-
-    def test_evaluate_circuit_hash_numeric(self):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        a = 0.3
-        b = 0.2
-
-        def circuit1():
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2():
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_symbolic(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_numeric_and_symbolic(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "a,b",
-        zip(np.linspace(0.1, 2 * np.pi, 3), np.linspace(0, 2 * np.pi, 3)),
-    )
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 0, 3), np.linspace(-2 * np.pi, 0, 3)),
-    )
-    def test_evaluate_circuit_hash_symbolic_bound_arguments_do_not_matter(self, a, b, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        def circuit1(a, b):
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([a, b], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    def test_evaluate_circuit_hash_numeric_tensor_return(self):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        a = 0.3
-        b = 0.2
-
-        def circuit1():
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2():
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_symbolic(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_numeric_and_symbolic_different_parameter(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.5, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_tensor_return(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_numeric_and_symbolic_return_type_does_not_matter(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.var(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        def circuit3(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.sample(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node3 = BaseQNode(circuit1, dev)
-        node3.evaluate([x, y], {})
-        circuit_hash_3 = node3.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2 == circuit_hash_3
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_hermitian(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        matrix = np.array([[1, 0], [0, 1]])
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.Hermitian(matrix, wires=[0]) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.Hermitian(matrix, wires=[0]) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 == circuit_hash_2
-
-class TestQNodeCircuitHashDifferentHashIntegration:
-    """Tests for checking that different circuit graph hashes are being created for different circuits in a QNode during evaluation (inside of _construct)"""
-
-    def test_evaluate_circuit_hash_numeric_different(self):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        a = 0.3
-        b = 0.2
-
-        def circuit1():
-            qml.RX(a, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        c = 0.6
-
-        def circuit2():
-            qml.RX(c, wires=[0])
-            qml.RY(b, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    def test_evaluate_circuit_hash_numeric_different_operation(self):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=2)
-
-        a = 0.3
-
-        def circuit1():
-            qml.RX(a, wires=[0])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2():
-            qml.RY(a, wires=[0])
-            return qml.expval(qml.PauliZ(0))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_numeric_and_symbolic_operation_differs(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RZ(y, wires=[1])          # <-------------------------------------- RZ
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])          # <-------------------------------------- RY
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_different_return_observable_vs_tensor(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.RZ(0.3, wires=[2])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic_different_order(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, 0.3, y, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 7), np.linspace(-2 * np.pi, 2 * np.pi, 7) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic_different_argument(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.5, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 2), np.linspace(-2 * np.pi, 2 * np.pi, 2) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic_different_wires(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])                               #<------ wires = [0, 1]
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[1, 0])                               #<------ wires = [1, 0]
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 2), np.linspace(-2 * np.pi, 2 * np.pi, 2) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_same_operation_has_numeric_and_symbolic_different_wires_in_return(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(1))         # <----- (0) @ (1)
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0) @ qml.PauliX(2))         # <----- (0) @ (2)
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
-    @pytest.mark.parametrize(
-        "x,y",
-        zip(np.linspace(-2 * np.pi, 2 * np.pi, 2), np.linspace(-2 * np.pi, 2 * np.pi, 2) ** 2 / 11),
-    )
-    def test_evaluate_circuit_hash_hermitian_different_matrices(self, x, y):
-        """Tests that the circuit hash of identical circuits are equal"""
-        dev = qml.device("default.qubit", wires=3)
-
-        matrix_1 = np.array([[1, 0], [0, 1]])
-        matrix_2 = np.array([[1, 0], [0, -1]])
-
-        def circuit1(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.Hermitian(matrix_1, wires=[0]) @ qml.PauliX(1))
-
-        node1 = BaseQNode(circuit1, dev)
-        node1.evaluate([x, y], {})
-        circuit_hash_1 = node1.circuit.circuit_hash
-
-        def circuit2(x, y):
-            qml.Rot(x, y, 0.3, wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.Hermitian(matrix_2, wires=[0]) @ qml.PauliX(1))
-
-        node2 = BaseQNode(circuit2, dev)
-        node2.evaluate([x, y], {})
-        circuit_hash_2 = node2.circuit.circuit_hash
-
-        assert circuit_hash_1 != circuit_hash_2
-
 class TestDecomposition:
     """Test for queue decomposition"""
